core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "elobs-word-updater"
version = "0.1.0"
description = "Fetch data from OpenStudyBuilder API and update Word document content controls"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "python-docx>=1.1",
    "lxml>=5.2",
    "docxcompose>=1.4",
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-html>=4.0",
    "httpx[cli]",
    "fastapi>=0.111",
    "uvicorn>=0.29",
]
# OAuth 2.0 support — not yet active, install when implementing auth
oauth = [
    "msal>=1.28",
]

[project.scripts]
elobs-word-updater = "elobs_word_updater.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
